requirements.txt
numpy==1.21.4
pandas==1.3.4
matplotlib==3.5.0
seaborn==0.11.2
streamlit==1.2.0
psycopg2==2.9.2
tabula-py==2.3.0